+++ reme/steps/cookbook/auto_fin/__init__.py
@@ -1,5 +1,6 @@
 """Cached news-case research workflow."""
 
-from .pipeline import AutoFinAnalysisStep, AutoFinDataStep, AutoFinPipelineStep
+from .analysis import AutoFinAnalysisStep, AutoFinPipelineStep
+from .data import AutoFinDataStep
 
 __all__ = ["AutoFinAnalysisStep", "AutoFinDataStep", "AutoFinPipelineStep"]
